% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggheatmap.R
\name{cluster_order}
\alias{cluster_order}
\title{Cluster Order}
\usage{
cluster_order(x, row, column, value, dist_method = "euclidean",
  hclust_method = "complete")
}
\arguments{
\item{x}{dataframe}

\item{row}{bare name of column to be the rows}

\item{column}{bare name of column to be the columns}

\item{value}{bare name of column to order by}

\item{dist_method}{distance method, passed to stats::dist}

\item{hclust_method}{hierchical clustering method, passed to stats::hclust}
}
\description{
Order columns in a dataframe by a value using hierchical clustering.  Used primarily when
making a heatmap style plot so the function takes two columns, one intended to the be rows
of the heatmap and the other intended to be the column.
}
