% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnetggplot.R
\name{cnetggplot}
\alias{cnetggplot}
\title{cnetggplot}
\usage{
cnetggplot(res, dds_res, algorithm = "kk", exclude = NULL)
}
\arguments{
\item{res}{A clusterProfiler results object}

\item{dds_res}{DESeq results cleaned up with biobroom::tidy}

\item{algorithm}{the layout algorithm for igraph}

\item{exclude}{any pathways to exclude from the path}
}
\description{
This function makes a similar plot to the 'cnetplot' in the Bioconductor package 'clusterProfiler but does it with ggplot using the ggraph package
}
