% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq_helpers.R
\name{compute_spia}
\alias{compute_spia}
\title{compute spia with provided gene list does the bulk of the work for run_spia}
\usage{
compute_spia(de, all, organism, database, cutoff)
}
\description{
compute spia with provided gene list does the bulk of the work for run_spia
}
