% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rna-seq_functions.R
\name{convertIDs}
\alias{convertIDs}
\title{Convert between gene IDs}
\usage{
convertIDs(ids, from, to, db, ifMultiple = c("putNA", "useFirst"))
}
\description{
from \url{http://www.bioconductor.org/help/course-materials/2014/SeattleOct2014/B02.1.1_RNASeqLab.html}
}

