% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rna-seq_functions.R
\name{create_gene_list}
\alias{create_gene_list}
\title{Create a gene list from a DESeq results object}
\usage{
create_gene_list(results, db = NULL, cutoff = 0.05, ncbi_link = FALSE)
}
\arguments{
\item{results}{DESeq results object}

\item{db}{An optional AnnotationDB object for gene identifiers}

\item{cutoff}{Cutoff for the adjusted p-value (or p-value if there is no adjusted). Defaults to 0.05}

\item{ncbi_link}{Logical, include NCBI link in the results. Defaults to FALSE.}
}
\description{
This is a function that takes a DESeqResults object and a data frame gene list
with only the genes that have and adjusted p-valu above cutoff.  Can include a link
to NCBI (inspired by the Reporting tools package)
}

