% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/microbiome_functions.R
\name{expand_swatch}
\alias{expand_swatch}
\title{Expand ggthemr swatch}
\usage{
scale_colour_manual(values = expand_swatch(num_cols))
}
\arguments{
\item{num_cols}{integer The number of colours you would like to create}
}
\value{
A vector of colours
}
\description{
Utility function to interpolate a ggthemr swatch to create more colours. Uses the swatch from the
currently activated theme.
}

