% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiome_helpers.R
\name{get_richness}
\alias{get_richness}
\title{Get a dataframe with richness estimates}
\usage{
get_richness(physeq, ...)
}
\arguments{
\item{physeq}{A phyloseq object}

\item{...}{Arguments passed to \code{estimate_richness}}
}
\value{
A dataframe with richness estimates and sample data
}
\description{
Wrapper around \code{\link{estimate_richness}} to get an easier to use dataframe
}
