% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiome_helpers.R
\name{get_sig_taxa}
\alias{get_sig_taxa}
\title{Get a list of significant taxa from a DESeq analysis}
\usage{
get_sig_taxa(dds, physeq, cutoff = 0.05, ...)
}
\arguments{
\item{dds}{valid DESeq2 object}

\item{physeq}{valid phyloseq object}

\item{cutoff}{pvalue cutoff}

\item{...}{other parameters passed to \code{\link{DESeq2::results}}}
}
\description{
Get a list of significant taxa from a DESeq analysis
}

