% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggheatmap.R
\name{ggheatmap}
\alias{ggheatmap}
\title{Make a ggplot heatmap}
\usage{
ggheatmap(x, row, column, value, scale = c("row", "column"),
  filter_empty = TRUE, dist_method = "euclidean",
  hclust_method = "complete", scico_pal = "vik")
}
\arguments{
\item{x}{dataframe}

\item{row}{bare name of column to be the rows}

\item{column}{bare name of column to be the columns}

\item{value}{bare name of column with values to fill the cells and order the rows and columns}

\item{dist_method}{distance method, passed to stats::dist}

\item{hclust_method}{hierchical clustering method, passed to stats::hclust}

\item{scico_pal}{colour palette to use from the package scico}
}
\value{
a ggplot object
}
\description{
Function to create a ggplot heatmap where the rows and columns are ordered with
heirarchical clustering, as in a traditional heatmap
}
