% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/microbiome_functions.R
\name{lfc_plot}
\alias{lfc_plot}
\title{Plot log-fold changes}
\usage{
lfc_plot(sigs_obj, order = "lfc", xtitle = "", colour = "Phylum")
}
\arguments{
\item{sigs_obj}{Dataframe with DESeq results}

\item{order}{string How to order the plot.  One of "lfc" or "Phylum".  Defaults to "lfc"}

\item{xtitle}{string Title of the x-axis. Defaults to "".}

\item{colour}{string Variable to colour the points by.  Must be a column in sigs_obj. Defaults to "Phylum"}
}
\value{
A ggplot object
}
\description{
Plot log-fold changes for taxa in a data frame that contains DESeq results
Can be used with output of \code{get_sig_taxa}
}

