% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/make_heatmap.R
\name{make_heatmap}
\alias{make_heatmap}
\title{Make a heatmap of a list of genes from a DESeq analysis}
\usage{
make_heatmap(rld, genes, anno_var, subset_samples = NULL, title = "", ...)
}
\arguments{
\item{rld}{A summarized experiment object from DESeq that has been transformed with rlog}

\item{genes}{A character vector of genes to plot.  Must match gene names in the rld object.}

\item{anno_var}{The metadat variable to annotate the columns with.  Must be a column in rld
column data}

\item{subset_samples}{A character vector of samples to subset the data.  Names must match the rld
object}

\item{title}{A plot title}

\item{...}{Other arguments passed to Heatmap}
}
\description{
Make a heatmap of a list of genes from a DESeq analysis
}

