% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiome_helpers.R
\name{make_ordination_plots}
\alias{make_ordination_plots}
\title{Make ordination plots}
\usage{
make_ordination_plots(physeq, ord_methods = c("NMDS", "PCoA"),
  distance = "bray", colour_var = NULL)
}
\arguments{
\item{physeq}{(Required). A phyloseq object}

\item{ord_methods}{A character vector containing the names of the ordinations
to plot. Default is \code{c("NMDS", "PCoA")}.}

\item{distance}{A character string. Default is "bray".  Name of the distance
method to use.  Value is passed to \code{\link{ordinate}}.}

\item{colour_var}{A character string containing the name of the variable to
use for colouring the plot.  Must be a valid column of
\code{sample_data(physeq)}}
}
\value{
Returns a \code{ggplot} object with the plots of interest
}
\description{
Function to quickly plot a variety of oridation methods Uses
\code{\link{plot_ordination}} from phyloseq
}

