% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalance_filtering.R
\name{make_prev_df}
\alias{make_prev_df}
\title{Make a prevalance dataframe}
\usage{
make_prev_df(physeq, prev_def)
}
\arguments{
\item{physeq}{a valid phyloseq object}

\item{prev_def}{prevalance definition, the minimum count for an OTU to defined
as present}
}
\description{
Function that takes a phyloseq object and creates a dataframe indicating
the prevalance of each OTU
}
