% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiome_helpers.R
\name{make_taxa_colours}
\alias{make_taxa_colours}
\title{Make colour palettes for taxa}
\usage{
make_taxa_colours(rank, physeq, palette = "iwanthue")
}
\arguments{
\item{rank}{Taxonomic rank to use}

\item{physeq}{valid phyloseq object}

\item{palette}{Either a valid RColorBrewer palette or "iwanthue" to generate
a random palette}
}
\description{
Make colour palettes for taxa
}
