% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mattsUtils.R
\docType{package}
\name{mattsUtils}
\alias{mattsUtils}
\alias{mattsUtils-package}
\title{mattsUtils: Convience functions and utilies to help with my analysis}
\description{
mattsUtils: Convience functions and utilies to help with my analysis
}

