% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{mypalettes}
\alias{mypalettes}
\title{Colour Palettes}
\format{\preformatted{List of 14
 $ intensePal1   : chr [1:14] "#105B1F" "#F845B3" "#4082D9" "#8A2829" ...
 $ fancyPal1     : chr [1:14] "#A6D9C6" "#DEDA5E" "#F0A4B5" "#6CE4A4" ...
 $ large22       : chr [1:22] "#75CADB" "#D1693C" "#87DB49" "#D089D3" ...
 $ large23       : chr [1:23] "#DA6056" "#6EDB5C" "#82B5D8" "#E2B93D" ...
 $ large37       : chr [1:37] "#F680A0" "#7BF12D" "#26C4BC" "#C48001" ...
 $ large14       : chr [1:14] "#95D997" "#CD8ACE" "#CB6F3F" "#C2D744" ...
 $ pal5          : chr [1:5] "#D33606" "#1762B1" "#36720F" "#2BB3A7" ...
 $ pal4          : chr [1:4] "#FF6730" "#1A87C3" "#096F36" "#B31589"
 $ pal12         : chr [1:12] "#D50869" "#68B978" "#3362CA" "#F67419" ...
 $ fancy5        : chr [1:5] "#CDBCDC" "#E0CF61" "#7DDCCA" "#EDA884" ...
 $ fancy4        : chr [1:4] "#4DC28E" "#F496D3" "#E5C848" "#08C9E8"
 $ intense4      : chr [1:4] "#BB4A53" "#58955B" "#966BB3" "#AD8031"
 $ intense14     : chr [1:14] "#6D3F86" "#6EAC38" "#CE4F34" "#67A2BD" ...
 $ taxa_colour_30: chr [1:30] "#C36768" "#77D265" "#7997BA" "#C9CA49" ...
}}
\usage{
mypalettes
}
\description{
generated by \href{http://tools.medialab.sciences-po.fr/iwanthue/}{iwanthue}
}
\keyword{datasets}

