% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rna-seq_functions.R
\name{plot_de_genes}
\alias{plot_de_genes}
\title{Plot differentially expressed genes
This function takes a DESeqResults object and contrasts argument and plots
the differentially expressed genes at padj cutoff}
\usage{
plot_de_genes(dds, contrast, cutoff = 0.05, ...)
}
\description{
Plot differentially expressed genes
This function takes a DESeqResults object and contrasts argument and plots
the differentially expressed genes at padj cutoff
}

