% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_diagnostics.R
\name{plot_diagnositics}
\alias{plot_diagnositics}
\title{Diagnostic plots for a DESeq differential expression analysis}
\usage{
plot_diagnositics(res, cutoff = 0.05, show_title = TRUE)
}
\arguments{
\item{res}{A DESeq results object}

\item{cutoff}{The p value to use as a cutoff for significantly differentially
expressed genes.  Defaults to 0.05}

\item{show_title}{If specified the description of the results object will be
printed as the plot's title}
}
\description{
Given a valid DESeq results object this function will make a p-value
histogram and a volcano plot.  The p-value histogram is constructed with the
p-values prior to multiple-testing correction and is used to diagnose if
something went wrong with your testing procedure.  See
\url{http://varianceexplained.org/statistics/interpreting-pvalue-histogram/}
for more details on interpreting the p-value histogram.  The volcano plot
gives you an idea of how many genes are differentially expressed in both
directions compared to their expression levels.
}

