% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq_helpers.R
\name{plot_featureCount_stats}
\alias{plot_featureCount_stats}
\title{Plot stats from featureCounts output}
\usage{
plot_featureCount_stats(stats, palette = "Set3")
}
\arguments{
\item{stats}{The count summary data}

\item{palette}{A valid RColorBrewer palette}
}
\description{
Takes the count summary data from a \code{\link[Rsubread]{featureCounts}}
object. Alternitavley you can load the counts.txt.summary file as a dataframe
and provide this if using the command line version of featureCounts.
}
