% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiome_helpers.R
\name{plot_OTUs}
\alias{plot_OTUs}
\title{Plot OTUs}
\usage{
plot_OTUs(physeq, otus, xaxis, fill, labeller = "Genus", scales = "free_y",
  palette = "Set1", glom = NULL, dds = NULL, justDf = FALSE)
}
\arguments{
\item{physeq}{a valid phyloseq object}

\item{otus}{a character vector of OTU ids, present in physeq.  Can be "all" to plot all OTUs}

\item{xaxis}{character, the sample column to plot on the x-axis}

\item{fill}{character, the sample column to use for colouring the boxes}

\item{labeller}{character, taxonomy rank to label the OTUs with}

\item{scales}{fixed or free scales, passed to facet_wrap}

\item{palette}{RColorBrewer palette to use}

\item{glom}{character, the taxonomic rank to glom at}

\item{dds}{valid DESeq2 object.  If present the normalized count data will be used instead of the raw count value.}

\item{justDf}{logical, should just the data be returned instead of plotting}
}
\description{
Makes a box plot from a list of OTUs
}
