% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rna-seq_functions.R
\name{plot_variable_genes}
\alias{plot_variable_genes}
\title{Plot most variable genes}
\usage{
plot_variable_genes(rld_obj, annotation_vars = NULL, top = 100,
  onlyGenes = FALSE, ...)
}
\description{
This function takes a DESeq results object that has been transformed (rlog)
and plots the top variable genes.  Also mean centers the data
}

