% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rna-seq_functions.R
\name{plotGenes}
\alias{plotGenes}
\title{Plot genes}
\usage{
plotGenes(geneList, dds, intgroup, db, title = NULL, plotType = c("box",
  "point", "justData"))
}
\description{
A function to plot normalized counts of genes from a DESeq object
}

