% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq_helpers.R
\name{plotPCA2}
\alias{plotPCA2}
\title{Plot PCA (modified)}
\usage{
plotPCA2(x, intgroup = "condition", ntop = 500, returnData = FALSE)
}
\description{
A slight modification from the DESeq function \code{\link[DESeq2]{plotPCA}}
to return the 3 PC also
}

