% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalance_filtering.R
\name{prev_filt}
\alias{prev_filt}
\title{Prevalance filter}
\usage{
prev_filt(physeq, prev_def, cutoff)
}
\arguments{
\item{physeq}{a valid phyloseq object}

\item{prev_def}{prevalance definition, the minimum count for an OTU to defined
as present}

\item{cutoff}{the proportion of samples that an OTU must be present in to be
included}
}
\value{
a phyloseq object with OTUs below the cutoff filtered out
}
\description{
Filter OTUs base on prevalance, defined as the number of samples the OTU is
considered 'present'
}
