% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalance_filtering.R
\name{prev_plot}
\alias{prev_plot}
\title{Prevelance Plot}
\usage{
prev_plot(physeq, cutoff = 0.1, prev_def = 1, phylum = NULL)
}
\arguments{
\item{physeq}{a valid phyloseq object}

\item{cutoff}{the proportion of samples that an OTU must be present in to be
included}

\item{prev_def}{prevalance definition, the minimum count for an OTU to defined
as present}

\item{phylum}{phylum/phyla to subset}
}
\value{
ggplot object
}
\description{
Plot the prevalance of each OTU, faceted by phyla
}
