% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiome_helpers.R
\name{replace_counts}
\alias{replace_counts}
\title{Replace counts}
\usage{
replace_counts(physeq, dds)
}
\description{
Replace OTU counts in a phyloseq object with the normalized counts from a DESeq object containing
the same OTUS
}
