% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq_helpers.R
\name{run_spia}
\alias{run_spia}
\title{Prepare gene lists and run spia}
\usage{
run_spia(res, dds, organism, database = c("reactome", "kegg"),
  key = NULL, cutoff = 0.05)
}
\arguments{
\item{res}{data frame with the columns 'estimate' and 'gene.  Usually DESeq
results processed with \code{\link{biobroom::tidy}}}

\item{dds}{valid DESeq object from which the results were generated}

\item{organism}{character, the organism to use. See available organisms with
\code{pathwayDatabases()}}

\item{database}{character, the pathway database to use, one of c("reactome",
"kegg").}

\item{key}{named character vector to use for converting gene IDs to Entrez
IDs if needed}

\item{cutoff}{numeric, p-value cutoff}
}
\value{
same as \code{\link{runSPIA}} but cleaned up, with database links
  added and pathways filtered to the privided cutoff
}
\description{
Wrapper for \code{\link{graphite::runSPIA}}. Will save the database file in
the current working directory if not already present.  Does some prep work
with gene ids and cleans up the output a bit.  Only supports reactome and
kegg for now.
}
