% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiome_helpers.R
\name{subset_samples_safe}
\alias{subset_samples_safe}
\title{Subset samples 2}
\usage{
subset_samples_safe(physeq, var, value)
}
\arguments{
\item{physeq}{a valid phyloseq object}

\item{var}{variable (column name) of sample data}

\item{value}{filter to keep only this value of the variable}
}
\description{
A programming safe version of \code{\link{phyloseq::subset_samples}}
}
