% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiome_helpers.R
\name{summarize_taxonomy}
\alias{summarize_taxonomy}
\title{Summarize taxonomy}
\usage{
summarize_taxonomy(physeq, grouping_tax = "Phylum",
  summary_tax = "Genus", filter = 0.01)
}
\arguments{
\item{physeq}{a valid phyloseq object}

\item{grouping_tax}{the higher level taxonomy to group within}

\item{summary_tax}{the lower level taxonomy to summarize}

\item{filter}{numeric, filter out members of the summary taxonomy below this fraction}
}
\value{
a data frame
}
\description{
Summarize a specified taxonomy within a higher level, usually phylum
}
