% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_phyloseq.R
\name{tidy_phyloseq}
\alias{tidy_phyloseq}
\title{Tidy phyloseq}
\usage{
tidy_phyloseq(physeq)
}
\arguments{
\item{physeq, }{a valid phyloseq object, sample names must not include
'Sample' or 'OTU' or any of the taxonomic ranks (kingdom,  phylum, etc)}
}
\value{
a tidy tibble
}
\description{
A tidy and fast version of phyloseq::psmelt, albeit with less checking
}
