% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_and_share.R
\name{upload_and_share}
\alias{upload_and_share}
\title{Upload and share}
\usage{
upload_and_share(file, path, remove_existing = TRUE)
}
\arguments{
\item{file}{local file to upload}

\item{path}{path on google drive to upload to}

\item{remove_existing}{logical, remove any existing files with the same name in the given path}
}
\value{
a list with the viewing and download link
}
\description{
Upload and share (view only) a file on google drive.  Will authenticate in
your browser the first time you use it
}
