% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiome_helpers.R
\name{vst_transform}
\alias{vst_transform}
\title{VST Tranform a phyloseq object}
\usage{
vst_transform(physeq)
}
\arguments{
\item{physeq}{a phyloseq object where the OTU table is raw counts}
}
\value{
a phyloseq object with vst normalized OTU table
}
\description{
Uses the \code{\link{getVarianceStabilizedData}} function from DESeq2 to vst
transform an OTU table
}
