% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rna-seq_functions.R
\name{convertIDs}
\alias{convertIDs}
\title{Convert between gene IDs}
\usage{
convertIDs(ids, from, to, db, ifMultiple = c("putNA", "useFirst"))
}
\arguments{
\item{ids}{A vector containing the gene IDs to convert}

\item{from}{Character, convert from}

\item{to}{Character, convert to}

\item{db}{A valid AnnotationDB object}

\item{ifMultiple}{What do if there are mutiple matches when converting}
}
\description{
from \url{http://www.bioconductor.org/help/course-materials/2014/SeattleOct2014/B02.1.1_RNASeqLab.html}
}

