% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/microbiome_functions.R
\name{get_sig_taxa}
\alias{get_sig_taxa}
\title{Get significant taxa}
\usage{
get_sig_taxa(res, alpha, physeq)
}
\arguments{
\item{res}{DESeq results object obtained from calling \code{results} on a DESeq object}

\item{alpha}{P-value cutoff}

\item{physeq}{The original phyloseq object}
}
\value{
A dataframe with the significant taxa
}
\description{
Get significantly differentially abundant taxa from a DESeq results object
}

