% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiome_helpers.R
\name{lfc_plot}
\alias{lfc_plot}
\title{Log fold-change plot}
\usage{
lfc_plot(x, tax_level = "Genus", colour = "Phylum", errors = FALSE)
}
\arguments{
\item{x}{results from DESeq2 differential analysis, processed with biobroom}

\item{tax_level}{the taxonomy level to group the y-axis by}

\item{colour}{character, the taxonomy level to colour the points}

\item{errors}{logical, should the error bars be included, defaults to FALSE}
}
\description{
Create plot that shows the log-fold change for various taxa
}
