% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiome_helpers.R
\name{plot_bar2}
\alias{plot_bar2}
\title{Plot bar 2}
\usage{
plot_bar2(physeq, rank = "Phylum", x = "Sample", xlabs = NULL)
}
\description{
An alternate version of \code{\link{phyloseq::plot_bar}}
}

