% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiome_helpers.R
\name{plot_genus_by_phylum}
\alias{plot_genus_by_phylum}
\title{Plot genus by phylum}
\usage{
plot_genus_by_phylum(physeq, phylum, filter = NULL, facet_by_fam = FALSE,
  xlabel = NULL)
}
\description{
Plot genus by phylum
}

