% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiome_helpers.R
\name{plot_OTUs}
\alias{plot_OTUs}
\title{Plot OTUs}
\usage{
plot_OTUs(physeq, otus, xaxis, fill, labeller = "Genus", scales = "free_y",
  palette = "Set1", glom = NULL, dds = NULL, justDf = FALSE,
  y_scale = c("counts", "log_counts", "relative"))
}
\arguments{
\item{physeq}{a valid phyloseq object with raw OTU counts that have not been transformed.}

\item{otus}{a character vector of OTU ids, present in physeq.  Can be "all" to plot all OTUs}

\item{xaxis}{character, the sample column to plot on the x-axis}

\item{fill}{character, the sample column to use for colouring the boxes}

\item{labeller}{character, taxonomy rank to label the OTUs with}

\item{scales}{fixed or free scales, passed to facet_wrap}

\item{palette}{RColorBrewer palette to use}

\item{glom}{character, the taxonomic rank to glom at}

\item{dds}{valid DESeq2 object.  If present the normalized count data will be used instead of the
raw count value.}

\item{justDf}{logical, should just the data be returned instead of plotting}

\item{y_scale}{character, the abundance scale on the y-axis.  Counts are just the raw OTU counts
unless a DESeq object is provided with the dds argument, in which case the normalized counts
will be used.  log_counts will transform the counts by log2 using a pseudcount of +0.5.
relative will use the relative abundance values instead of counts.}
}
\description{
Makes a box plot from a list of OTUs
}
