% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rna-seq_functions.R
\name{plotPCA2}
\alias{plotPCA2}
\title{Plot PCA (modified)}
\usage{
plotPCA2(x, intgroup = "condition", ntop = 500, returnData = FALSE)
}
\description{
A slight modification from the DESeq function \code{\link[DESeq2]{plotPCA}} to return the 3 PC also
}

